Improve tests and add release test

+++ tests/test.bats
@@ -1,24 +1,34 @@
 setup() {
   export DIR="$( cd "$( dirname "$BATS_TEST_FILENAME" )" >/dev/null 2>&1 && pwd )/.."
-  export TESTDIR=$(mktemp -d -t testmemcached-XXXXXXXXXX)
+  export TESTDIR=~/tmp/testelasticsearch
+  mkdir -p $TESTDIR
   export PROJNAME=testmemcached
   export DDEV_NON_INTERACTIVE=true
   ddev delete -Oy ${PROJNAME} || true
   cd "${TESTDIR}"
-  ddev config --project-name=${PROJNAME} --project-type=drupal9 --docroot=web --create-docroot
+  ddev config --project-name=${PROJNAME}
   ddev start
 }
 
 teardown() {
-  cd ${TESTDIR}
+  cd ${TESTDIR} || ( printf "unable to cd to ${TESTDIR}\n" && exit 1 )
   ddev delete -Oy ${DDEV_SITENAME}
-  rm -rf ${TESTDIR}
+  [ "${TESTDIR}" != "" ] && rm -rf ${TESTDIR}
 }
 
-@test "basic installation" {
-  cd ${TESTDIR}
+@test "install from directory" {
+  cd ${TESTDIR} || ( printf "unable to cd to ${TESTDIR}\n" && exit 1 )
+  echo "# ddev get ${DIR} with project ${PROJNAME} in ${TESTDIR} ($(pwd))" >&3
   ddev get ${DIR}
   ddev restart
   v=$(ddev exec 'printf "version\nquit\nquit\n" | nc memcached 11211')
   [[ "${v}" = VERSION* ]]
+}
+
+@test "install from release" {
+  cd ${TESTDIR} || ( printf "unable to cd to ${TESTDIR}\n" && exit 1 )
+  echo "# ddev get drud/ddev-elasticsearch with project ${PROJNAME} in ${TESTDIR} ($(pwd))" >&3
+  ddev get drud/ddev-elasticsearch
+  ddev restart
+  ddev exec "curl -s elasticsearch:9200" | grep "${PROJNAME}-elasticsearch"
 }
